feature(rslint_parser): Synthesized Tokens
#1867 proposes to convert any node to an `Unknown*` node if any required child is missing, not present in the source code.

This allows to type the field accessors for required children as `T` rather than `Result<T>`, which is a nice ergonomics improvement.

However, this would result in many nodes turning into `Unknown` nodes if only a single token is missing. For example, `if (test) {` turns into an `Unknown` node because the closing `}` is missing.

At this point, the parser can be fairly certain that the user is about to complete the `if` statement, thus, it would be safe to assume the closing `}` is present in the source code.

This PR adds support for synthesized tokens. A synthesized token is a token that isn't present in the original source text (isn't written) but has been inserted by the parser to form a valid tree.

Inserting synthesized punctuation tokens should be safe as most analyzers will just be skipping over them. The only use case where checking for the source presence is relevant right now is the formatter where we don't want to insert not written tokens.